Rename format files to formatter.
