+++ core/containers/array.h
@@ -2,7 +2,7 @@
 
 #include "core/defines.h"
 #include "core/assert.h"
-#include "core/format.h"
+#include "core/formatter.h"
 #include "core/memory/memory.h"
 #include "core/serialization/serializer.h"
