+++ core/containers/hash_table.h
@@ -3,7 +3,7 @@
 #include "core/defines.h"
 #include "core/hash.h"
 #include "core/utils.h"
-#include "core/format.h"
+#include "core/formatter.h"
 #include "core/memory/memory.h"
 #include "core/containers/array.h"
 #include "core/serialization/serializer.h"
