+++ core/containers/string.h
@@ -4,7 +4,7 @@
 #include "core/assert.h"
 #include "core/defer.h"
 #include "core/hash.h"
-#include "core/format.h"
+#include "core/formatter.h"
 #include "core/memory/memory.h"
 #include "core/containers/array.h"
 #include "core/serialization/serializer.h"
