+++ core/formatter.h
