+++ core/platform/platform_win32.cpp
@@ -2,7 +2,7 @@
 
 #include "core/defer.h"
 #include "core/assert.h"
-#include "core/format.h"
+#include "core/formatter.h"
 #include "core/logger.h"
 #include "core/memory/memory.h"
 #include "core/containers/array.h"
