+++ core/result.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "core/defines.h"
-#include "core/format.h"
+#include "core/formatter.h"
 #include "core/containers/string.h"
 
 struct Error
